+++ lib/rubocop/cop/style/alias.rb
@@ -8,9 +8,21 @@ module Rubocop
       class Alias < Cop
         MSG = 'Use alias_method instead of alias.'
 
-        # TODO: Make this check context aware - alias_method is not
-        #   available outside of classes/modules.
+        def on_block(node)
+          method, _args, body = *node
+
+          _receiver, method_name = *method
+
+          # using alias is the only option in certain scenarios
+          # in such scenarios we don't want to report an offence
+          if method_name == :instance_exec
+            on_node(:alias, body) { |n| ignore_node(n) }
+          end
+        end
+
         def on_alias(node)
+          return if ignored_node?(node)
+
           # alias_method can't be used with global variables
           new, old = *node